% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_caseBoxes.R
\name{mod_caseBoxes}
\alias{mod_caseBoxes}
\alias{mod_caseBoxes_ui}
\alias{mod_caseBoxes_server}
\title{caseBoxes UI Function}
\usage{
mod_caseBoxes_ui(id)

mod_caseBoxes_server(input, output, session, counts)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{counts}{Reactive expression yielding the named vector of cases by type.}
}
\description{
A shiny Module for displaying the number by type as colored boxes.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    tagList(
      babyMeasure:::mod_caseBoxes_ui("boxes")
    )
  )
  server <- function(input, output, session) {
    counts <- reactive({
      invalidateLater(2000)
      stats::setNames(sample.int(1000, 3), c("today", "week", "month"))
    })
    callModule(Covid19:::mod_caseBoxes_server, "boxes", counts = counts)
  }
  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
}

}
\keyword{internal}
